+++ script/tool/lib/src/build_examples_command.dart
@@ -9,7 +9,10 @@ import 'package:file/file.dart';
 import 'package:path/path.dart' as p;
 import 'package:platform/platform.dart';
 
-import 'common.dart';
+import 'common/core.dart';
+import 'common/plugin_command.dart';
+import 'common/plugin_utils.dart';
+import 'common/process_runner.dart';
 
 /// Key for IPA.
 const String kIpa = 'ipa';
