+++ script/tool/lib/src/publish_plugin_command.dart
@@ -14,7 +14,10 @@ import 'package:pub_semver/pub_semver.dart';
 import 'package:pubspec_parse/pubspec_parse.dart';
 import 'package:yaml/yaml.dart';
 
-import 'common.dart';
+import 'common/core.dart';
+import 'common/git_version_finder.dart';
+import 'common/plugin_command.dart';
+import 'common/process_runner.dart';
 
 @immutable
 class _RemoteInfo {
