+++ script/tool/test/common/plugin_command_test.dart
@@ -2,24 +2,20 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:async';
-import 'dart:convert';
 import 'dart:io';
 
 import 'package:args/command_runner.dart';
 import 'package:file/file.dart';
 import 'package:file/memory.dart';
-import 'package:flutter_plugin_tools/src/common.dart';
+import 'package:flutter_plugin_tools/src/common/plugin_command.dart';
+import 'package:flutter_plugin_tools/src/common/process_runner.dart';
 import 'package:git/git.dart';
-import 'package:http/http.dart' as http;
-import 'package:http/testing.dart';
 import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
-import 'package:pub_semver/pub_semver.dart';
 import 'package:test/test.dart';
 
-import 'common_test.mocks.dart';
-import 'util.dart';
+import '../util.dart';
+import 'plugin_command_test.mocks.dart';
 
 @GenerateMocks(<Type>[GitDir])
 void main() {
@@ -360,355 +356,6 @@ packages/plugin3/plugin3.dart
 
         expect(plugins, unorderedEquals(<String>[plugin1.path]));
       });
-    });
-  });
-
-  group('$GitVersionFinder', () {
-    late FileSystem fileSystem;
-    late List<List<String>?> gitDirCommands;
-    late String gitDiffResponse;
-    String? mergeBaseResponse;
-    late MockGitDir gitDir;
-
-    setUp(() {
-      fileSystem = MemoryFileSystem();
-      createPackagesDirectory(fileSystem: fileSystem);
-      gitDirCommands = <List<String>?>[];
-      gitDiffResponse = '';
-      gitDir = MockGitDir();
-      when(gitDir.runCommand(any, throwOnError: anyNamed('throwOnError')))
-          .thenAnswer((Invocation invocation) {
-        gitDirCommands.add(invocation.positionalArguments[0] as List<String>?);
-        final MockProcessResult mockProcessResult = MockProcessResult();
-        if (invocation.positionalArguments[0][0] == 'diff') {
-          when<String?>(mockProcessResult.stdout as String?)
-              .thenReturn(gitDiffResponse);
-        } else if (invocation.positionalArguments[0][0] == 'merge-base') {
-          when<String?>(mockProcessResult.stdout as String?)
-              .thenReturn(mergeBaseResponse);
-        }
-        return Future<ProcessResult>.value(mockProcessResult);
-      });
-      processRunner = RecordingProcessRunner();
-    });
-
-    test('No git diff should result no files changed', () async {
-      final GitVersionFinder finder = GitVersionFinder(gitDir, 'some base sha');
-      final List<String> changedFiles = await finder.getChangedFiles();
-
-      expect(changedFiles, isEmpty);
-    });
-
-    test('get correct files changed based on git diff', () async {
-      gitDiffResponse = '''
-file1/file1.cc
-file2/file2.cc
-''';
-      final GitVersionFinder finder = GitVersionFinder(gitDir, 'some base sha');
-      final List<String> changedFiles = await finder.getChangedFiles();
-
-      expect(
-          changedFiles, equals(<String>['file1/file1.cc', 'file2/file2.cc']));
-    });
-
-    test('get correct pubspec change based on git diff', () async {
-      gitDiffResponse = '''
-file1/pubspec.yaml
-file2/file2.cc
-''';
-      final GitVersionFinder finder = GitVersionFinder(gitDir, 'some base sha');
-      final List<String> changedFiles = await finder.getChangedPubSpecs();
-
-      expect(changedFiles, equals(<String>['file1/pubspec.yaml']));
-    });
-
-    test('use correct base sha if not specified', () async {
-      mergeBaseResponse = 'shaqwiueroaaidf12312jnadf123nd';
-      gitDiffResponse = '''
-file1/pubspec.yaml
-file2/file2.cc
-''';
-
-      final GitVersionFinder finder = GitVersionFinder(gitDir, null);
-      await finder.getChangedFiles();
-      verify(gitDir.runCommand(
-          <String>['diff', '--name-only', mergeBaseResponse!, 'HEAD']));
-    });
-
-    test('use correct base sha if specified', () async {
-      const String customBaseSha = 'aklsjdcaskf12312';
-      gitDiffResponse = '''
-file1/pubspec.yaml
-file2/file2.cc
-''';
-      final GitVersionFinder finder = GitVersionFinder(gitDir, customBaseSha);
-      await finder.getChangedFiles();
-      verify(gitDir
-          .runCommand(<String>['diff', '--name-only', customBaseSha, 'HEAD']));
-    });
-  });
-
-  group('$PubVersionFinder', () {
-    test('Package does not exist.', () async {
-      final MockClient mockClient = MockClient((http.Request request) async {
-        return http.Response('', 404);
-      });
-      final PubVersionFinder finder = PubVersionFinder(httpClient: mockClient);
-      final PubVersionFinderResponse response =
-          await finder.getPackageVersion(package: 'some_package');
-
-      expect(response.versions, isEmpty);
-      expect(response.result, PubVersionFinderResult.noPackageFound);
-      expect(response.httpResponse.statusCode, 404);
-      expect(response.httpResponse.body, '');
-    });
-
-    test('HTTP error when getting versions from pub', () async {
-      final MockClient mockClient = MockClient((http.Request request) async {
-        return http.Response('', 400);
-      });
-      final PubVersionFinder finder = PubVersionFinder(httpClient: mockClient);
-      final PubVersionFinderResponse response =
-          await finder.getPackageVersion(package: 'some_package');
-
-      expect(response.versions, isEmpty);
-      expect(response.result, PubVersionFinderResult.fail);
-      expect(response.httpResponse.statusCode, 400);
-      expect(response.httpResponse.body, '');
-    });
-
-    test('Get a correct list of versions when http response is OK.', () async {
-      const Map<String, dynamic> httpResponse = <String, dynamic>{
-        'name': 'some_package',
-        'versions': <String>[
-          '0.0.1',
-          '0.0.2',
-          '0.0.2+2',
-          '0.1.1',
-          '0.0.1+1',
-          '0.1.0',
-          '0.2.0',
-          '0.1.0+1',
-          '0.0.2+1',
-          '2.0.0',
-          '1.2.0',
-          '1.0.0',
-        ],
-      };
-      final MockClient mockClient = MockClient((http.Request request) async {
-        return http.Response(json.encode(httpResponse), 200);
-      });
-      final PubVersionFinder finder = PubVersionFinder(httpClient: mockClient);
-      final PubVersionFinderResponse response =
-          await finder.getPackageVersion(package: 'some_package');
-
-      expect(response.versions, <Version>[
-        Version.parse('2.0.0'),
-        Version.parse('1.2.0'),
-        Version.parse('1.0.0'),
-        Version.parse('0.2.0'),
-        Version.parse('0.1.1'),
-        Version.parse('0.1.0+1'),
-        Version.parse('0.1.0'),
-        Version.parse('0.0.2+2'),
-        Version.parse('0.0.2+1'),
-        Version.parse('0.0.2'),
-        Version.parse('0.0.1+1'),
-        Version.parse('0.0.1'),
-      ]);
-      expect(response.result, PubVersionFinderResult.success);
-      expect(response.httpResponse.statusCode, 200);
-      expect(response.httpResponse.body, json.encode(httpResponse));
-    });
-  });
-
-  group('pluginSupportsPlatform', () {
-    test('no platforms', () async {
-      final Directory plugin = createFakePlugin('plugin', packagesDir);
-
-      expect(pluginSupportsPlatform('android', plugin), isFalse);
-      expect(pluginSupportsPlatform('ios', plugin), isFalse);
-      expect(pluginSupportsPlatform('linux', plugin), isFalse);
-      expect(pluginSupportsPlatform('macos', plugin), isFalse);
-      expect(pluginSupportsPlatform('web', plugin), isFalse);
-      expect(pluginSupportsPlatform('windows', plugin), isFalse);
-    });
-
-    test('all platforms', () async {
-      final Directory plugin = createFakePlugin(
-        'plugin',
-        packagesDir,
-        isAndroidPlugin: true,
-        isIosPlugin: true,
-        isLinuxPlugin: true,
-        isMacOsPlugin: true,
-        isWebPlugin: true,
-        isWindowsPlugin: true,
-      );
-
-      expect(pluginSupportsPlatform('android', plugin), isTrue);
-      expect(pluginSupportsPlatform('ios', plugin), isTrue);
-      expect(pluginSupportsPlatform('linux', plugin), isTrue);
-      expect(pluginSupportsPlatform('macos', plugin), isTrue);
-      expect(pluginSupportsPlatform('web', plugin), isTrue);
-      expect(pluginSupportsPlatform('windows', plugin), isTrue);
-    });
-
-    test('some platforms', () async {
-      final Directory plugin = createFakePlugin(
-        'plugin',
-        packagesDir,
-        isAndroidPlugin: true,
-        isIosPlugin: false,
-        isLinuxPlugin: true,
-        isMacOsPlugin: false,
-        isWebPlugin: true,
-        isWindowsPlugin: false,
-      );
-
-      expect(pluginSupportsPlatform('android', plugin), isTrue);
-      expect(pluginSupportsPlatform('ios', plugin), isFalse);
-      expect(pluginSupportsPlatform('linux', plugin), isTrue);
-      expect(pluginSupportsPlatform('macos', plugin), isFalse);
-      expect(pluginSupportsPlatform('web', plugin), isTrue);
-      expect(pluginSupportsPlatform('windows', plugin), isFalse);
-    });
-
-    test('inline plugins are only detected as inline', () async {
-      // createFakePlugin makes non-federated pubspec entries.
-      final Directory plugin = createFakePlugin(
-        'plugin',
-        packagesDir,
-        isAndroidPlugin: true,
-        isIosPlugin: true,
-        isLinuxPlugin: true,
-        isMacOsPlugin: true,
-        isWebPlugin: true,
-        isWindowsPlugin: true,
-      );
-
-      expect(
-          pluginSupportsPlatform('android', plugin,
-              requiredMode: PlatformSupport.inline),
-          isTrue);
-      expect(
-          pluginSupportsPlatform('android', plugin,
-              requiredMode: PlatformSupport.federated),
-          isFalse);
-      expect(
-          pluginSupportsPlatform('ios', plugin,
-              requiredMode: PlatformSupport.inline),
-          isTrue);
-      expect(
-          pluginSupportsPlatform('ios', plugin,
-              requiredMode: PlatformSupport.federated),
-          isFalse);
-      expect(
-          pluginSupportsPlatform('linux', plugin,
-              requiredMode: PlatformSupport.inline),
-          isTrue);
-      expect(
-          pluginSupportsPlatform('linux', plugin,
-              requiredMode: PlatformSupport.federated),
-          isFalse);
-      expect(
-          pluginSupportsPlatform('macos', plugin,
-              requiredMode: PlatformSupport.inline),
-          isTrue);
-      expect(
-          pluginSupportsPlatform('macos', plugin,
-              requiredMode: PlatformSupport.federated),
-          isFalse);
-      expect(
-          pluginSupportsPlatform('web', plugin,
-              requiredMode: PlatformSupport.inline),
-          isTrue);
-      expect(
-          pluginSupportsPlatform('web', plugin,
-              requiredMode: PlatformSupport.federated),
-          isFalse);
-      expect(
-          pluginSupportsPlatform('windows', plugin,
-              requiredMode: PlatformSupport.inline),
-          isTrue);
-      expect(
-          pluginSupportsPlatform('windows', plugin,
-              requiredMode: PlatformSupport.federated),
-          isFalse);
-    });
-
-    test('federated plugins are only detected as federated', () async {
-      const String pluginName = 'plugin';
-      final Directory plugin = createFakePlugin(
-        pluginName,
-        packagesDir,
-        isAndroidPlugin: true,
-        isIosPlugin: true,
-        isLinuxPlugin: true,
-        isMacOsPlugin: true,
-        isWebPlugin: true,
-        isWindowsPlugin: true,
-      );
-
-      createFakePubspec(
-        plugin,
-        name: pluginName,
-        androidSupport: PlatformSupport.federated,
-        iosSupport: PlatformSupport.federated,
-        linuxSupport: PlatformSupport.federated,
-        macosSupport: PlatformSupport.federated,
-        webSupport: PlatformSupport.federated,
-        windowsSupport: PlatformSupport.federated,
-      );
-
-      expect(
-          pluginSupportsPlatform('android', plugin,
-              requiredMode: PlatformSupport.federated),
-          isTrue);
-      expect(
-          pluginSupportsPlatform('android', plugin,
-              requiredMode: PlatformSupport.inline),
-          isFalse);
-      expect(
-          pluginSupportsPlatform('ios', plugin,
-              requiredMode: PlatformSupport.federated),
-          isTrue);
-      expect(
-          pluginSupportsPlatform('ios', plugin,
-              requiredMode: PlatformSupport.inline),
-          isFalse);
-      expect(
-          pluginSupportsPlatform('linux', plugin,
-              requiredMode: PlatformSupport.federated),
-          isTrue);
-      expect(
-          pluginSupportsPlatform('linux', plugin,
-              requiredMode: PlatformSupport.inline),
-          isFalse);
-      expect(
-          pluginSupportsPlatform('macos', plugin,
-              requiredMode: PlatformSupport.federated),
-          isTrue);
-      expect(
-          pluginSupportsPlatform('macos', plugin,
-              requiredMode: PlatformSupport.inline),
-          isFalse);
-      expect(
-          pluginSupportsPlatform('web', plugin,
-              requiredMode: PlatformSupport.federated),
-          isTrue);
-      expect(
-          pluginSupportsPlatform('web', plugin,
-              requiredMode: PlatformSupport.inline),
-          isFalse);
-      expect(
-          pluginSupportsPlatform('windows', plugin,
-              requiredMode: PlatformSupport.federated),
-          isTrue);
-      expect(
-          pluginSupportsPlatform('windows', plugin,
-              requiredMode: PlatformSupport.inline),
-          isFalse);
     });
   });
 }
