+++ script/tool/test/common/plugin_command_test.mocks.dart
